+++ portlets/ams-portlet/docroot/WEB-INF/src/com/liferay/ams/service/persistence/AssetPersistenceImpl.java
@@ -241,7 +241,9 @@ public class AssetPersistenceImpl extends BasePersistenceImpl<Asset>
 						asset.getPrimaryKeyObj());
 			}
 
-			session.delete(asset);
+			if (asset != null) {
+				session.delete(asset);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -250,7 +252,9 @@ public class AssetPersistenceImpl extends BasePersistenceImpl<Asset>
 			closeSession(session);
 		}
 
-		clearCache(asset);
+		if (asset != null) {
+			clearCache(asset);
+		}
 
 		return asset;
 	}
