+++ portlets/chat-portlet/docroot/WEB-INF/src/com/liferay/chat/service/persistence/StatusPersistenceImpl.java
@@ -323,7 +323,9 @@ public class StatusPersistenceImpl extends BasePersistenceImpl<Status>
 						status.getPrimaryKeyObj());
 			}
 
-			session.delete(status);
+			if (status != null) {
+				session.delete(status);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -332,7 +334,9 @@ public class StatusPersistenceImpl extends BasePersistenceImpl<Status>
 			closeSession(session);
 		}
 
-		clearCache(status);
+		if (status != null) {
+			clearCache(status);
+		}
 
 		return status;
 	}
