+++ portlets/contacts-portlet/docroot/WEB-INF/src/com/liferay/contacts/service/persistence/EntryPersistenceImpl.java
@@ -293,7 +293,9 @@ public class EntryPersistenceImpl extends BasePersistenceImpl<Entry>
 						entry.getPrimaryKeyObj());
 			}
 
-			session.delete(entry);
+			if (entry != null) {
+				session.delete(entry);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -302,7 +304,9 @@ public class EntryPersistenceImpl extends BasePersistenceImpl<Entry>
 			closeSession(session);
 		}
 
-		clearCache(entry);
+		if (entry != null) {
+			clearCache(entry);
+		}
 
 		return entry;
 	}
