+++ portlets/knowledge-base-portlet/docroot/WEB-INF/src/com/liferay/knowledgebase/service/persistence/KBArticlePersistenceImpl.java
@@ -898,7 +898,9 @@ public class KBArticlePersistenceImpl extends BasePersistenceImpl<KBArticle>
 						kbArticle.getPrimaryKeyObj());
 			}
 
-			session.delete(kbArticle);
+			if (kbArticle != null) {
+				session.delete(kbArticle);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -907,7 +909,9 @@ public class KBArticlePersistenceImpl extends BasePersistenceImpl<KBArticle>
 			closeSession(session);
 		}
 
-		clearCache(kbArticle);
+		if (kbArticle != null) {
+			clearCache(kbArticle);
+		}
 
 		return kbArticle;
 	}
