+++ portlets/mail-portlet/docroot/WEB-INF/src/com/liferay/mail/service/persistence/AccountPersistenceImpl.java
@@ -287,7 +287,9 @@ public class AccountPersistenceImpl extends BasePersistenceImpl<Account>
 						account.getPrimaryKeyObj());
 			}
 
-			session.delete(account);
+			if (account != null) {
+				session.delete(account);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -296,7 +298,9 @@ public class AccountPersistenceImpl extends BasePersistenceImpl<Account>
 			closeSession(session);
 		}
 
-		clearCache(account);
+		if (account != null) {
+			clearCache(account);
+		}
 
 		return account;
 	}
