+++ portlets/mail-portlet/docroot/WEB-INF/src/com/liferay/mail/service/persistence/AttachmentPersistenceImpl.java
@@ -266,7 +266,9 @@ public class AttachmentPersistenceImpl extends BasePersistenceImpl<Attachment>
 						attachment.getPrimaryKeyObj());
 			}
 
-			session.delete(attachment);
+			if (attachment != null) {
+				session.delete(attachment);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -275,7 +277,9 @@ public class AttachmentPersistenceImpl extends BasePersistenceImpl<Attachment>
 			closeSession(session);
 		}
 
-		clearCache(attachment);
+		if (attachment != null) {
+			clearCache(attachment);
+		}
 
 		return attachment;
 	}
