+++ portlets/mail-portlet/docroot/WEB-INF/src/com/liferay/mail/service/persistence/MessagePersistenceImpl.java
@@ -311,7 +311,9 @@ public class MessagePersistenceImpl extends BasePersistenceImpl<Message>
 						message.getPrimaryKeyObj());
 			}
 
-			session.delete(message);
+			if (message != null) {
+				session.delete(message);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -320,7 +322,9 @@ public class MessagePersistenceImpl extends BasePersistenceImpl<Message>
 			closeSession(session);
 		}
 
-		clearCache(message);
+		if (message != null) {
+			clearCache(message);
+		}
 
 		return message;
 	}
