+++ portlets/opensocial-portlet/docroot/WEB-INF/src/com/liferay/opensocial/service/persistence/OAuthTokenPersistenceImpl.java
@@ -315,7 +315,9 @@ public class OAuthTokenPersistenceImpl extends BasePersistenceImpl<OAuthToken>
 						oAuthToken.getPrimaryKeyObj());
 			}
 
-			session.delete(oAuthToken);
+			if (oAuthToken != null) {
+				session.delete(oAuthToken);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -324,7 +326,9 @@ public class OAuthTokenPersistenceImpl extends BasePersistenceImpl<OAuthToken>
 			closeSession(session);
 		}
 
-		clearCache(oAuthToken);
+		if (oAuthToken != null) {
+			clearCache(oAuthToken);
+		}
 
 		return oAuthToken;
 	}
