+++ portlets/private-messaging-portlet/docroot/WEB-INF/src/com/liferay/privatemessaging/service/persistence/UserThreadPersistenceImpl.java
@@ -359,7 +359,9 @@ public class UserThreadPersistenceImpl extends BasePersistenceImpl<UserThread>
 						userThread.getPrimaryKeyObj());
 			}
 
-			session.delete(userThread);
+			if (userThread != null) {
+				session.delete(userThread);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -368,7 +370,9 @@ public class UserThreadPersistenceImpl extends BasePersistenceImpl<UserThread>
 			closeSession(session);
 		}
 
-		clearCache(userThread);
+		if (userThread != null) {
+			clearCache(userThread);
+		}
 
 		return userThread;
 	}
