+++ portlets/social-coding-portlet/docroot/WEB-INF/src/com/liferay/socialcoding/service/persistence/JIRAChangeItemPersistenceImpl.java
@@ -271,7 +271,9 @@ public class JIRAChangeItemPersistenceImpl extends BasePersistenceImpl<JIRAChang
 						jiraChangeItem.getPrimaryKeyObj());
 			}
 
-			session.delete(jiraChangeItem);
+			if (jiraChangeItem != null) {
+				session.delete(jiraChangeItem);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -280,7 +282,9 @@ public class JIRAChangeItemPersistenceImpl extends BasePersistenceImpl<JIRAChang
 			closeSession(session);
 		}
 
-		clearCache(jiraChangeItem);
+		if (jiraChangeItem != null) {
+			clearCache(jiraChangeItem);
+		}
 
 		return jiraChangeItem;
 	}
