+++ portlets/social-coding-portlet/docroot/WEB-INF/src/com/liferay/socialcoding/service/persistence/JIRAIssuePersistenceImpl.java
@@ -479,7 +479,9 @@ public class JIRAIssuePersistenceImpl extends BasePersistenceImpl<JIRAIssue>
 						jiraIssue.getPrimaryKeyObj());
 			}
 
-			session.delete(jiraIssue);
+			if (jiraIssue != null) {
+				session.delete(jiraIssue);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -488,7 +490,9 @@ public class JIRAIssuePersistenceImpl extends BasePersistenceImpl<JIRAIssue>
 			closeSession(session);
 		}
 
-		clearCache(jiraIssue);
+		if (jiraIssue != null) {
+			clearCache(jiraIssue);
+		}
 
 		return jiraIssue;
 	}
