+++ portlets/test-pacl-portlet/docroot/WEB-INF/src/com/liferay/testpacl/service/persistence/FooPersistenceImpl.java
@@ -241,7 +241,9 @@ public class FooPersistenceImpl extends BasePersistenceImpl<Foo>
 				foo = (Foo)session.get(FooImpl.class, foo.getPrimaryKeyObj());
 			}
 
-			session.delete(foo);
+			if (foo != null) {
+				session.delete(foo);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -250,7 +252,9 @@ public class FooPersistenceImpl extends BasePersistenceImpl<Foo>
 			closeSession(session);
 		}
 
-		clearCache(foo);
+		if (foo != null) {
+			clearCache(foo);
+		}
 
 		return foo;
 	}
